[SHARED] Move execCmd from fs to own folder
