+++ packages/common/src/execCmd/createFullCmd.js
